requirements.txt
# JAC Language Runtime
jaclang>=0.9.3

# FastAPI and ASGI
fastapi==0.124.0
uvicorn[standard]==0.38.0
starlette==0.50.0

# Database
sqlalchemy==2.0.44
alembic==1.17.2
aiosqlite>=0.19.0

# Database Drivers (commented out - uncomment for production)
# psycopg2-binary==2.9.11      # PostgreSQL
# mysql-connector-python>=8.2.0  # MySQL
# pymongo>=4.6.0               # MongoDB

redis==7.1.0
neo4j==6.0.3

# Authentication and Security
PyJWT==2.9.0
cryptography==41.0.8
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Data Validation and Serialization
pydantic==2.12.5
pydantic-settings==2.0.3
marshmallow==3.20.0

# HTTP Client
httpx==0.27.2
requests==2.31.0
aiohttp==3.8.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Environment Management
python-dotenv==1.2.1
click==8.1.0

# Development and Testing
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-cov==4.1.0
black==23.7.0
isort==5.13.2
flake8==6.1.0

# Monitoring and Logging
prometheus-client==0.17.0
structlog==23.1.0

# Utilities
python-multipart==0.0.9
email-validator==2.1.0
pytz==2023.3
python-dateutil==2.8.2

# AI/ML (for future integration)
openai==1.12.0
transformers==4.36.0
torch==2.1.2

# Performance
gunicorn==21.2.0
slowapi==0.1.9